Fix agones bug

+++ backend/src/game-server/game-server.service.ts
@@ -23,7 +23,6 @@ export class GameServerService implements OnModuleInit, OnModuleDestroy {
         private moduleRef: ModuleRef,
         private authService: AuthService,
         private schedulerRegistry: SchedulerRegistry,
-        private agones: AgonesSDK,
         private playerService: PlayersService,
         private messagingService: FirebaseMessagingService,
         private gamesService: GamesService,
@@ -31,6 +30,7 @@ export class GameServerService implements OnModuleInit, OnModuleDestroy {
 
     private logger: Logger = new Logger(GameServerService.name);
     private server: Server;
+    private agones: AgonesSDK = null;
     // Current game;
     private game: GameDocument = null;
     // If shutdown by user
@@ -43,6 +43,7 @@ export class GameServerService implements OnModuleInit, OnModuleDestroy {
         this.logger.log('Loaded server instance');
         // Agones
         if (process.env.AGONES_ENABLED === 'true') {
+            this.agones = this.moduleRef.get(AgonesSDK);
             this.agonesEnabled = true;
             this.setupAgones();
         }
@@ -102,42 +103,63 @@ export class GameServerService implements OnModuleInit, OnModuleDestroy {
     }
 
     async handleLogin(client: Socket) {
-        const user = await this.authService.validateUser(
-            (client.handshake.auth as any).token,
-        );
-        // If invalid credentials
-        if (!user) {
-            client.emit('disconnectLogin', { reason: 'Incorrect credentials' });
+        if (this.game === null || this.game.state === GameState.CREATED) {
+            this.logger.log('Client disconnected: Server not ready');
+            client.emit('message', {
+                disconnected: true,
+                message: 'Server not ready',
+            });
             client.disconnect();
         } else {
-            // Store users details on the server
-            this.playerService.addPlayerRaw(client, user);
-            // Send welcome message
-            this.schedulerRegistry.addTimeout(
-                `game-server:hello:${user.uid}`,
-                setTimeout(() => {
-                    this.messagingService.sendToDevice(
-                        user.devices.map((device) => device.token),
-                        {
-                            notification: {
-                                title: 'QuizKan',
-                                body: 'Welcome to The Game ',
-                            },
-                        },
-                    );
-                }, 5000),
+            const user = await this.authService.validateUser(
+                (client.handshake.auth as any).token,
             );
-            // Send login success
-            client.emit('loginSuccess', {
-                user: user.toJSON(),
-                time: dayjs().toISOString(),
-            });
-            this.logger.log(`User ${user.uid} logged in`);
+            // If invalid credentials
+            if (!user) {
+                client.emit('message', {
+                    disconnected: true,
+                    message: 'Incorrect credentials',
+                });
+                client.disconnect();
+            } else if (this.playerService.getPlayerById(user.uid)) {
+                client.emit('message', {
+                    disconnected: true,
+                    message: 'You are already logged in from somewhere!',
+                });
+                client.disconnect();
+            } else {
+                // Store users details on the server
+                this.playerService.addPlayerRaw(client, user);
+                // Send welcome message
+                this.schedulerRegistry.addTimeout(
+                    `game-server:hello:${user.uid}`,
+                    setTimeout(() => {
+                        this.messagingService.sendToDevice(
+                            user.devices.map((device) => device.token),
+                            {
+                                notification: {
+                                    title: 'QuizKan',
+                                    body: 'Welcome to The Game ',
+                                },
+                            },
+                        );
+                    }, 5000),
+                );
+                // Send login success
+                client.emit('loginSuccess', {
+                    user: user.toJSON(),
+                    time: dayjs().toISOString(),
+                });
+                this.logger.log(`User ${user.uid} logged in`);
+            }
         }
     }
 
     async handleDisconnect(client: Socket) {
         // Remove users details on the server
+        if (this.playerService.getPlayerBySocket(client) == null) {
+            return;
+        }
         const player = this.playerService.getPlayerBySocket(client);
         this.playerService.removePlayer(player);
     }
